+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to Shiny for Python will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### New features
+
+* Added `session.allow_reconnect()` method to enable, disable, or _force_ client reconnection behavior after disconnection. (#2102)
+
 ## [1.5.0] - 2025-09-11
 
 ### New features